[#1] SharedPhotoCell V,VM 분리 및 업로드 로직 수정
aynsc / awiat로 코드 변경
